+++ test/components/test_utils.py
@@ -5,7 +5,7 @@ import geopandas.testing
 import ohsome
 import pytest
 import shapely
-from climatoology.base.exception import ClimatoologyUserError
+from climatoology.base.exception import ClimatoologyUserError, InputValidationError
 from ohsome import OhsomeClient
 from ohsome.exceptions import OhsomeException
 from ohsome_filter_to_sql.main import validate_filter
@@ -25,7 +25,7 @@ def test_check_paths_count_limit(default_aoi, expected_compute_input, responses_
         )
 
     # test false situation
-    with pytest.raises(ClimatoologyUserError):
+    with pytest.raises(InputValidationError):
         check_paths_count_limit(default_aoi, OhsomeClient(), 5000)